BREAKING CHANGE: Move ServerAnalytics to Facades namespace.
