+++ tests/RequestDetailsTest.php
@@ -4,7 +4,7 @@ namespace OhSeeSoftware\LaravelServerAnalytics\Tests;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use OhSeeSoftware\LaravelServerAnalytics\RequestDetails;
-use OhSeeSoftware\LaravelServerAnalytics\ServerAnalytics;
+use OhSeeSoftware\LaravelServerAnalytics\Facades\ServerAnalytics;
 
 class RequestDetailsTest extends TestCase
 {
